requirements.txt
# 1. Создайте этот requirements.txt
streamlit==1.28.0
pandas==1.5.3
numpy==1.24.3
plotly==5.17.0
openai==1.6.1
python-telegram-bot==20.7
python-dotenv==1.0.0
matplotlib==3.8.0
seaborn==0.12.2
requests==2.31.0
reportlab==4.0.4
markdown==3.5.1
xlsxwriter==3.1.2
aiohttp==3.9.1
pillow==10.1.0

